# zoho-middleware/openapi.yaml
openapi: 3.0.3
info:
  title: Steins & Vines Middleware API
  description: |
    Backend middleware for the Steins & Vines fermentation supply store.
    Bridges the frontend (GitHub Pages) with Zoho Inventory, Zoho Books,
    Zoho Bookings, and Global Payments.

    **Authentication model:**
    - Most `GET` endpoints are public (no key required).
    - Most `POST`/`PUT` endpoints require `X-Api-Key: <API_SECRET_KEY>`.
    - `POST /api/checkout` is public — protected by reCAPTCHA + rate limiting instead.
    - `POST /product-requests` and `POST /api/contact` are fully public.
    - Auth routes (`/auth/*`) are pre-auth and never require a key.

    **CORS:** Requests must originate from `steinsandvines.ca`,
    `staging.steinsandvines.ca`, `localhost:3001`, or `localhost:8080`.

    **Rate limits:** `/api/*` endpoints are limited to 60 req/min per IP.
    Payment endpoints (`/api/checkout`, `/api/payment/*`, `/api/pos/sale`,
    `/api/kiosk/sale`) share a stricter 10 req/min limiter.
  version: "1.0.0"
  contact:
    name: Steins & Vines IT
    email: hello@steinsandvines.ca
servers:
  - url: https://steinsandvines-middleware.railway.app
    description: Production (Railway)
  - url: http://localhost:3001
    description: Local development
...
components:
  securitySchemes:
    ApiKeyHeader:
      type: apiKey
      in: header
      name: X-Api-Key
      description: Must equal the `API_SECRET_KEY` Railway env var.
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: "Cart is empty"
...
    CheckoutResponse:
      type: object
      required: [ok, salesorder_number, deposit_amount, balance_due]
      properties:
        ok:
          type: boolean
          example: true
        salesorder_id:
          type: string
          nullable: true
          example: "109900000000345678"
        salesorder_number:
          type: string
          example: "SO-00042"
        deposit_amount:
          type: number
          format: double
          example: 50.00
        balance_due:
          type: number
          format: double
          example: 99.99

    # -------------------------------------------------------------------------
    # Bookings
    # -------------------------------------------------------------------------
...
paths:
...
  /api/checkout:
    post:
      tags: [checkout]
      summary: Submit a ferment-in-store reservation
      description: |
        Creates a Zoho Books Sales Order from a cart. The Zoho contact is
        always resolved server-side from the supplied email (lookup or create);
        a client-supplied `customer_id` is intentionally ignored.

        Prices are anchored to the authoritative catalog cache — client-supplied
        `rate` values are never used. The Makers Fee service item must be present
        in the cart alongside kit items.

        **Payment flow (v1.4.0+):**
        1. Supply `payment_token` — server charges the card server-side inside
           this endpoint, then creates the Zoho Sales Order.
        2. Supply `transaction_id` — legacy path where the frontend pre-charged.

        On Zoho failure after a successful charge, the transaction is
        automatically voided.

        This endpoint is **public** — not API-key protected. Instead it is
        protected by reCAPTCHA v3 (fail-open when unconfigured) and a rate
        limiter (10 req/min per IP).
      operationId: postCheckout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckoutRequest'
            example:
              customer:
                name: "Jane Smith"
                email: "jane@example.com"
                phone: "604-555-0100"
              items:
                - item_id: "109900000000123456"
                  name: "Chilliwack Chardonnay Wine Kit"
                  quantity: 1
                  rate: 149.99
                - item_id: "109900000000654321"
                  name: "Makers Fee"
                  quantity: 1
                  rate: 10.00
              notes: "First time making wine — please provide guidance."
              timeslot: "2026-03-20 10:00 AM"
              idempotency_key: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
      responses:
        "201":
          description: Sales order created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckoutResponse'
              example:
                ok: true
                salesorder_id: "109900000000345678"
                salesorder_number: "SO-00042"
                deposit_amount: 0
                balance_due: 159.99
        "400":
          description: Validation error (missing fields, unknown item, Makers Fee missing, etc.).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Zoho not authenticated (and route is not in offline-capable list).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "402":
          description: Card declined during server-side charge.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "409":
          description: Payment already processed (duplicate transaction_id replay detected).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "429":
          description: Rate limit exceeded (10 req/min per IP).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "502":
          description: Upstream error (Zoho or payment gateway).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "503":
          description: Pricing temporarily unavailable (catalog cache empty).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ===========================================================================
  # Bookings
  # ===========================================================================
...
  /api/bookings:
    post:
      tags: [bookings]
      summary: Create an appointment
      description: >
        Creates an appointment in Zoho Bookings. Returns `booking_id` and
        the formatted `timeslot` string for use as a Zoho Books custom field.

        If Zoho is offline, returns a placeholder booking_id prefixed with
        `PENDING-` so the checkout flow can continue; the full order notification
        is sent by `/api/checkout`.
      operationId: postBooking
      security:
        - ApiKeyHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingRequest'
            example:
              date: "2026-03-20"
              time: "10:00 AM"
              customer:
                name: "Jane Smith"
                email: "jane@example.com"
                phone: "604-555-0100"
              notes: "First time — interested in a wine kit."
      responses:
        "201":
          description: Appointment created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingResponse'
              example:
                ok: true
                booking_id: "109900000000789012"
                timeslot: "2026-03-20 10:00 AM"
        "400":
          description: Validation error (missing date/time, invalid email, etc.).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "403":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "502":
          description: Upstream Zoho Bookings error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ===========================================================================
  # Contacts
  # ===========================================================================
  /api/contacts:
    post:
      tags: [contacts]
      summary: Find or create a Zoho Books contact
      description: >
        Searches Zoho Books for an existing contact by email. If not found,
        creates a new customer contact. Returns the `contact_id` for use in
        downstream Sales Order creation.

        When Zoho is offline, returns `{ contact_id: "offline", offline: true }`
        so the checkout flow can continue.
      operationId: postContact
      security:
        - ApiKeyHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContactRequest'
            example:
              name: "Jane Smith"
              email: "jane@example.com"
              phone: "604-555-0100"
      responses:
        "200":
          description: Existing contact found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContactResponse'
        "201":
          description: New contact created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContactResponse'
              example:
                contact_id: "109900000000112233"
                created: true
        "400":
          description: Missing or invalid email.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "403":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "502":
          description: Upstream Zoho error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ===========================================================================
  # Payments
  # ===========================================================================
...
  /api/pos/sale:
    post:
      tags: [pos]
      summary: Push a legacy POS sale to the terminal
      description: >
        Sends a sale request to the GP UPA terminal via Meet in the Cloud.
        The terminal displays the amount and waits for card tap/insert/swipe.
        On success, creates a Zoho Books Invoice and payment record as a
        background operation (non-fatal). Returns immediately after terminal
        approval.

        Rate limit: 10 req/min per IP.
      operationId: postPosSale
      security:
        - ApiKeyHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PosSaleRequest'
            example:
              amount: 99.99
              salesorder_number: "SO-00042"
              customer_name: "Jane Smith"
      responses:
        "200":
          description: Terminal approved the payment.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PosSaleResponse'
        "400":
          description: Missing or invalid amount.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "402":
          description: Terminal declined the payment.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "403":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "429":
          description: Rate limit exceeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "502":
          description: Terminal error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "503":
          description: POS terminal not configured.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/kiosk/sale:
    post:
      tags: [pos]
      summary: Process a full kiosk sale
      description: |
        Processes a complete kiosk sale:
        1. Validates cart items against the kiosk products cache (price anchoring).
        2. Computes tax server-side from `KIOSK_TAX_RATE` (default 5%).
        3. Sends payment to the GP POS terminal.
        4. On success: creates a Zoho Books Invoice and records a payment.
        5. Invalidates the kiosk products cache.
        6. Returns receipt data.

        If invoice creation fails after payment, the terminal transaction is
        automatically voided.

        Client-supplied `rate` and `tax_total` are both **ignored**.

        Rate limit: 10 req/min per IP.
      operationId: postKioskSale
      security:
        - ApiKeyHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KioskSaleRequest'
            example:
              items:
                - item_id: "109900000000123456"
                  name: "Chilliwack Chardonnay Wine Kit"
                  quantity: 1
              reference_number: "KIOSK-001"
              idempotency_key: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
      responses:
        "201":
          description: Sale completed and invoice created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KioskSaleResponse'
              example:
                ok: true
                transaction_id: "gp-txn-abc123"
                auth_code: "123456"
                invoice_id: "109900000000567890"
                invoice_number: "INV-00123"
                reference_number: "KIOSK-001"
                subtotal: 149.99
                tax_total: 7.50
                total: 157.49
                date: "2026-03-12"
        "400":
          description: Validation error (empty cart, unknown item_id, etc.).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "402":
          description: Terminal declined the payment.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "403":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "429":
          description: Rate limit exceeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "502":
          description: Invoice creation failed after payment (transaction voided).
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  payment_voided:
                    type: boolean
                  voided_transaction_id:
                    type: string
        "503":
          description: POS terminal not configured or catalog cache unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "504":
          description: Terminal timed out (no transaction created).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/pos/collect:
    post:
      tags: [pos]
      summary: Send SO balance to Helcim terminal for collection
      description: >
        Called by Zoho Inventory custom button (Deluge invokeUrl).
        Fetches the Sales Order balance and pushes it to the Helcim
        Smart Terminal. Returns 202 immediately; webhook handles
        the payment result and updates Zoho.
      operationId: postPosCollect
      security:
        - ApiKeyHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [salesorder_id]
              properties:
                salesorder_id:
                  type: string
                  description: Zoho Sales Order ID
      responses:
        "202":
          description: Payment sent to terminal.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  salesorder_number:
                    type: string
                  amount:
                    type: number
                  status:
                    type: string
                    enum: [pending]
        "400":
          description: Invalid input, zero balance, or void/closed order.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Sales Order not found in Zoho.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "409":
          description: Payment already pending for this order.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "503":
          description: Helcim terminal not configured.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "502":
          description: Terminal or Zoho API error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...